+++ PyISY/Nodes/group.py
@@ -1,4 +1,6 @@
 """Representation of groups (scenes) from an ISY."""
+from VarEvents import Property
+
 from ..constants import PROTO_GROUP, VALUE_UNKNOWN
 from .nodebase import NodeBase
 
@@ -16,7 +18,7 @@ class Group(NodeBase):
     |  controllers: List of the controllers in this group.
     |  spoken: The string of the Notes Spoken field.
 
-    :ivar hasChildren: Boolean value indicating that group has no children.
+    :ivar has_children: Boolean value indicating that group has no children.
     :ivar members: List of the members of this group.
     :ivar controllers: List of the controllers of this group.
     :ivar name: The name of this group.