+++ PyISY/Nodes/node.py
@@ -44,7 +44,7 @@ class Node(NodeBase):
 
     :ivar status: A watched property that indicates the current status of the
                   node.
-    :ivar hasChildren: Property indicating that there are no more children.
+    :ivar has_children: Property indicating that there are no more children.
     """
 
     def __init__(
@@ -77,7 +77,7 @@ class Node(NodeBase):
         self.status.update(
             state.get(ATTR_VALUE, VALUE_UNKNOWN), force=True, silent=True
         )
-        self.controlEvents = EventEmitter()
+        self.control_events = EventEmitter()
         super().__init__(
             nodes, nid, name, family_id=family_id, aux_properties=aux_properties
         )