+++ PyISY/Variables/__init__.py
@@ -86,7 +86,7 @@ class Variables:
         """Parse the XML Variable Definitions from the ISY."""
         try:
             xmldocs = [minidom.parseString(xml) for xml in xmls]
-        except:
+        except (AttributeError, KeyError, ValueError, TypeError, IndexError):
             self.isy.log.error("%s: Variables", XML_PARSE_ERROR)
             return
 
@@ -101,7 +101,7 @@ class Variables:
         """Parse XML from the controller with details about the variables."""
         try:
             xmldoc = minidom.parseString(xml)
-        except:
+        except (AttributeError, KeyError, ValueError, TypeError, IndexError):
             self.isy.log.error("%s: Variables", XML_PARSE_ERROR)
             return