+++ PyISY/networking.py
@@ -63,7 +63,7 @@ class NetworkResources:
         """
         try:
             xmldoc = minidom.parseString(xml)
-        except:
+        except (AttributeError, KeyError, ValueError, TypeError, IndexError):
             self.isy.log.error("%s: NetworkResources", XML_PARSE_ERROR)
         else:
             features = xmldoc.getElementsByTagName(TAG_NET_RULE)
@@ -88,7 +88,7 @@ class NetworkResources:
         xml = self.isy.conn.get_network()
         self.parse(xml)
 
-    def updateThread(self):
+    def update_threaded(self):
         """
         Continually update the class until it is told to stop.
 
@@ -102,7 +102,7 @@ class NetworkResources:
         try:
             val = int(val)
             return self.get_by_id(val)
-        except:
+        except (ValueError, KeyError):
             return self.get_by_name(val)
 
     def __setitem__(self, val, value):
@@ -118,7 +118,7 @@ class NetworkResources:
         try:
             ind = self.nids.index(val)
             return self.get_by_index(ind)
-        except:
+        except (ValueError, KeyError):
             return None
 
     def get_by_name(self, val):
@@ -130,7 +130,7 @@ class NetworkResources:
         try:
             ind = self.nnames.index(val)
             return self.get_by_index(ind)
-        except:
+        except (ValueError, KeyError):
             return None
 
     def get_by_index(self, val):